+++ packages/react/src/index.ts
@@ -22,3 +22,7 @@ export * from "@sk-web-gui/divider";
 
 export * from "@sk-web-gui/user-menu";
 export * from "@sk-web-gui/table";
+
+export * from "@sk-web-gui/tag";
+
+export * from "@sk-web-gui/pagination";