+++ packages/accordion/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@0.1.4...@sk-web-gui/accordion@0.1.5) (2022-04-21)
+
+**Note:** Version bump only for package @sk-web-gui/accordion
+
+
+
+
+
+## [0.1.4](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@0.1.3...@sk-web-gui/accordion@0.1.4) (2022-04-13)
+
+**Note:** Version bump only for package @sk-web-gui/accordion
+
+
+
+
+
 ## [0.1.3](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/accordion@0.1.2...@sk-web-gui/accordion@0.1.3) (2022-03-31)
 
 **Note:** Version bump only for package @sk-web-gui/accordion