+++ packages/divider/CHANGELOG.md
@@ -3,11 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.2](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/divider@0.1.1...@sk-web-gui/divider@0.1.2) (2022-03-31)
+## [0.1.5](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/divider@0.1.4...@sk-web-gui/divider@0.1.5) (2022-04-21)
 
 **Note:** Version bump only for package @sk-web-gui/divider
 
-## 0.1.1 (2022-03-29)
+
+
+
+
+## [0.1.4](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/divider@0.1.3...@sk-web-gui/divider@0.1.4) (2022-04-13)
+
+**Note:** Version bump only for package @sk-web-gui/divider
+
+
+
+
+
+## [0.1.3](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/divider@0.1.2...@sk-web-gui/divider@0.1.3) (2022-04-05)
+
+**Note:** Version bump only for package @sk-web-gui/divider
+
+## [0.1.2](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/divider@0.1.1...@sk-web-gui/divider@0.1.2) (2022-03-31)
 
 **Note:** Version bump only for package @sk-web-gui/divider