+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/radio@0.1.2...@sk-web-gui/radio@0.1.3) (2022-05-02)
+
+**Note:** Version bump only for package @sk-web-gui/radio
+
+
+
+
+
+## [0.1.2](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/radio@0.1.1...@sk-web-gui/radio@0.1.2) (2022-04-21)
+
+**Note:** Version bump only for package @sk-web-gui/radio
+
+
+
+
+
 ## 0.1.1 (2022-02-18)
 
 **Note:** Version bump only for package @sk-web-gui/radio