+++ packages/react/CHANGELOG.md
@@ -3,11 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.6](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.5...@sk-web-gui/react@0.1.6) (2022-03-31)
+## [0.1.13](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.12...@sk-web-gui/react@0.1.13) (2022-05-04)
 
 **Note:** Version bump only for package @sk-web-gui/react
 
-## [0.1.5](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.4...@sk-web-gui/react@0.1.5) (2022-03-29)
+
+
+
+
+## [0.1.12](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.11...@sk-web-gui/react@0.1.12) (2022-05-04)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
+
+
+
+
+## [0.1.11](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.10...@sk-web-gui/react@0.1.11) (2022-05-02)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
+
+
+
+
+## [0.1.10](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.9...@sk-web-gui/react@0.1.10) (2022-04-21)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
+
+
+
+
+## [0.1.9](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.8...@sk-web-gui/react@0.1.9) (2022-04-14)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
+
+
+
+
+## [0.1.8](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.7...@sk-web-gui/react@0.1.8) (2022-04-13)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
+
+
+
+
+## [0.1.7](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.6...@sk-web-gui/react@0.1.7) (2022-04-05)
+
+**Note:** Version bump only for package @sk-web-gui/react
+
+## [0.1.6](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/react@0.1.5...@sk-web-gui/react@0.1.6) (2022-03-31)
 
 **Note:** Version bump only for package @sk-web-gui/react